==> 2017_07_cp_analysis/02_normalize_migration.Rmd <==
---
title: "load_data"
author: "TB"
date: "5/15/2017"
output: html_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
library("tidyverse")
library("magrittr")
library("knitr")
```

# Input 
Define the input folders and the csv files used to load the single cell data generated by cellprofiler.
```{r}
analysis_folder <- '../../../analysis/2017_morph_gradient_3/'
csv_tracks <- 'clean_tracks_cp.csv'
csv_normalized_migration <- 'normalized_migration.csv'
```

# Metadata 
```{r}
strata <- c('Metadata_condition','Metadata_dose','Metadata_id','Metadata_date','Metadata_matrix')
```

# Load data 
We load the data and "validate" the tracks. This includes
* removal of all small tracks (Track_Length)
* removal of tracks with more than one object per frame (Track_One_Cell)
```{r}
population <- read_csv(file.path(analysis_folder, csv_tracks)) %>%
  mutate(Metadata_condition = replace(Metadata_condition, Metadata_condition == "water", "vehicle")) %>%
  filter(Track_Length > 19) %>%
  filter(Track_One_Cell == 1)
  
```
Load tracking data created using MATLAB as normalized and not normalized data. 

The normalization was performed using cytominer::normalize and the following variables 
are normalized: 
* "Track_Speed",
* "Track_Speed_Y",
* "Track_Speed_X",
* "Track_xFMI",
* "Track_yFMI",
* "Track_Directionality",
* "Track_Distance_Traveled", 
* "Track_CI"

We define the variables for normalization and use cytominer::normalize
```{r}
feature_migration = c("Track_Speed","Track_Speed_Y","Track_Speed_X","Track_xFMI","Track_yFMI","Track_Directionality","Track_Distance_Traveled","Track_CI")

# optionally we can run 
#population <- cytominer::generalized_log(population, variables = feature_var)

normalized <-
  cytominer::normalize(
    population = population,
    variables = feature_migration,
    strata =  "Metadata_id",
    sample = population %>% filter(Metadata_condition == "vehicle"),
    operation = "robustize"
  )

normalized %<>% dplyr::collect() 

```

## Save normalized data.
```{r}
write_csv(normalized, file.path(analysis_folder, csv_normalized_migration))
```
