==> 2017_07_cp_analysis/03_normalize_features.Rmd <==
---
title: "load_data"
author: "TB"
date: "5/15/2017"
output: html_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
library("tidyverse")
library("magrittr")
library("knitr")
library("corrplot")
library("replyr")
library("stringr")
source("../utilities.R")
```

# Input 
Define the input folders and the csv files used to load the single cell data generated by cellprofiler.
```{r}
analysis_folder <- '../../../analysis/2017_morph_gradient_3/'
csv_tracks <- 'clean_tracks_cp.csv'
csv_cells <- 'single_cell_data.csv'
csv_normalized_features <- 'normalized_features.csv'
csv_collapsed_features <- 'collapsed_features.csv' 
```

# Metadata 
```{r}
strata <- c('Metadata_condition','Metadata_dose','Metadata_id','Metadata_date','Metadata_matrix')
```


# load csv files from all sub folders 
population <- data.frame()
for (iDir in list.dirs(analysis_folder)) {
  file_name <-  file.path(iDir, csv_cells)
  if (file.exists(file_name)) {
    population <- rbind(population,suppressMessages(read_csv(file_name)))
  }
}

population %<>% 
  mutate(Metadata_timePoint = Metadata_frame) %>%
  dplyr::mutate(Metadata_timePoint = as.double(Metadata_timePoint)) %>%
  mutate(Metadata_frame = Metadata_timePoint) %>%
  select(Metadata_id, Metadata_date, Metadata_matrix, Metadata_condition, Metadata_dose, Metadata_frame, everything())

write_csv(population, file.path(analysis_folder, 'single_cell_data.csv'))


## tracks
```{r}
tracks <- read_csv(file.path(analysis_folder, csv_tracks) )  %>% 
  mutate(Metadata_condition = replace(Metadata_condition, Metadata_condition == "water", "vehicle")) 

population <- read_csv(file.path(analysis_folder, csv_cells) )  %>% 
  mutate(Metadata_condition = replace(Metadata_condition, Metadata_condition == "water", "vehicle")) %>%
  select(-AreaShape_EulerNumber, -AreaShape_Center_X, -AreaShape_Center_Y, -AreaShape_MedianRadius)
```

# we only keep valid paths, i.e. cells corresponding to short tracks are removed  
```{r}
valid_tracks <- tracks %>% 
  filter(Track_Length > 19) %>% 
  filter(Track_One_Cell == 1) %>%
  select_(.dots = c(strata, "TrackObjects_Label_30")) 

population_valid <- semi_join(population, valid_tracks, by = c(strata, "TrackObjects_Label_30"))
```

# Compare complete cases for 
```{r}
sum(complete.cases(population) == TRUE) %>% print
sum(complete.cases(population_valid) == TRUE) %>% print

```

# Which columns has many NA values? 
which columns have the highest amount of na values? 
```{r}
na_per_column(population_valid) %>% 
  arrange(-na_per_column) %>% 
  print 
```
The columns TrackObjects_FinalAge_30 and TrackObjects_ParentImageNumber_30 are removed, the values in TrackObjects_Linearity_30 are set to zero
```{r}
population_valid %<>% 
  select(-TrackObjects_FinalAge_30,-TrackObjects_ParentImageNumber_30 )

population_valid$TrackObjects_Linearity_30[is.na(population_valid$TrackObjects_Linearity_30)] = 0
```

# How many complete cases do we have now? 
```{r}
sum(complete.cases(population_valid) == TRUE) %>%
  print
```

# Using our complete_cases all non complete caes are removed 
```{r}
population_valid %<>%
  filter(complete.cases(.))
```

# we update our features 
```{r}
feature_var <-
  colnames(population_valid) %>%
  str_subset("^Area|^Texture|^Track")

feature_var = feature_var[!feature_var %in% "TrackObjects_Label_30"]
```

## single cell data 
The output of CellProfiler track module is loaded as CSV, tracked using cytominer::track and 
stored in the data folder as "track.csv". 

# normalize features 
```{r}
normalized <-
  cytominer::normalize(
    population = population_valid,
    variables = feature_var,
    strata =  "Metadata_id",
    sample = population_valid %>% filter(Metadata_condition == "vehicle"),
    operation = "robustize"
  )

normalized %<>% dplyr::collect() 

write_csv(normalized, file.path(analysis_folder, csv_normalized_features))
```

# Check for na rows in normalized
```{r}
na_per_column(normalized) %>% arrange(-na_per_column) %>% print
sum(complete.cases(population_valid) == TRUE) / nrow(population_valid)
sum(complete.cases(normalized) == TRUE) / nrow(normalized)
```

## collapse features to single tracks 
```{r}
normalized %<>% group_by_(.dots = c(strata, "TrackObjects_Label_30"))

aggregated <- cytominer::aggregate(
    population = normalized,
    variables = feature_var,
    strata = c(strata, "TrackObjects_Label_30"),
    operation = "median"
  ) %>% collect()

write_csv(aggregated, file.path(analysis_folder, csv_collapsed_features))
```
