==> 2017_07_cp_analysis/04_audit.Rmd <==
---
title: "load_data"
author: "TB"
date: "5/15/2017"
output: html_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
library("tidyverse")
library("magrittr")
library("knitr")
library("corrplot")
library("stringr")
source("../utilities.R")
library("SNFtool")
```

# In this notebook we measure the quality of the tracking results using the same method as implemented in the audit.R function of cytominer_scripts repository. 


# Input 
Define the input folders and the csv files used to load the single cell data generated by cellprofiler.
```{r}
analysis_folder <- '../../../analysis/2017_morph_gradient_3/'
csv_in <- 'collapsed_features.csv'
```

# Metadata 
```{r}
strata <- c('Metadata_condition','Metadata_dose','Metadata_id','Metadata_date','Metadata_matrix')
```

# load data
```{r}
population <- read_csv(file.path(analysis_folder, csv_in) ) %>%
  select(-TrackObjects_Linearity_30 ) 

population %<>% 
  filter(complete.cases(.))
```
##
```{r}

group_by = c("Metadata_condition","Metadata_dose","Metadata_matrix")
```

## select Area features
```{r}
df <- population %>%
  select(-TrackObjects_Label_30, -TrackObjects_ParentObjectNumber_30) %>%
  select(-TrackObjects_TrajectoryX_30, -TrackObjects_TrajectoryY_30)

helper_string <- c("^Area","^Texture","^Track","^Texture|^Track","^Area|^Track","^Area|^Texture","^Area|^Texture|^Track")

result_cor <- tibble()
result_aff <- tibble()
seeds <- sample(1:1000, 5, replace = FALSE)
```

```{r}
for (seed in seeds) {
  set.seed(seed)
  for (feature_string in helper_string) {
    feature_var <-
    colnames(df) %>%
    str_subset(feature_string)
  
  feature_var = feature_var[!feature_var %in% "TrackObjects_Label_30"]
  
  result_cor <- df %>% 
    filter(Metadata_matrix == "FN") %>%
    fraction_strong(.,group_by = group_by, feature_var, method = "correlation")  %>%
    mutate(Metadata_matrix = "FN") %>%
    mutate(feature = feature_string) %>%
    mutate(seed = seed) %>%
    rbind(.,result_cor)
  
  result_cor <- df %>% 
    filter(Metadata_matrix == "HEM") %>%
    fraction_strong(.,group_by = group_by, feature_var) %>%
    mutate(Metadata_matrix = "HEM") %>%
    mutate(feature = feature_string) %>%
     mutate(seed = seed) %>%
    rbind(.,result_cor)
  }
}
```


```{r}
for (seed in seeds) {
  set.seed(seed)
  for (feature_string in helper_string) {
    feature_var <-
    colnames(df) %>%
    str_subset(feature_string)
  
  feature_var = feature_var[!feature_var %in% "TrackObjects_Label_30"]
  
  result_aff <- df %>% 
    filter(Metadata_matrix == "FN") %>%
    fraction_strong(.,group_by = group_by, feature_var, method = "affinity")  %>%
    mutate(Metadata_matrix = "FN") %>%
    mutate(feature = feature_string) %>%
    mutate(seed = seed) %>%
    rbind(.,result_aff)
  
  result_aff <- df %>% 
    filter(Metadata_matrix == "HEM") %>%
    fraction_strong(.,group_by = group_by, feature_var, method = "affinity") %>%
    mutate(Metadata_matrix = "HEM") %>%
    mutate(feature = feature_string) %>%
     mutate(seed = seed) %>%
    rbind(.,result_aff)
  }
}
```


```{r}
result_cor %>% 
  group_by(feature,Metadata_matrix) %>%
  select(-seed) %>%
  summarise_each(funs(mean)) %>%
  arrange(Metadata_matrix, fraction_strong) %>%
  knitr::kable()

result_aff %>% 
  group_by(feature,Metadata_matrix) %>%
  select(-seed) %>%
  summarise_each(funs(mean)) %>%
  arrange(Metadata_matrix, fraction_strong) %>%
  knitr::kable()
```

# Now we compare the signal when we add tracking feature 
## load tracking data 
```{r}
migration <- read_csv(file.path(analysis_folder, "normalized_migration.csv"))
population_2 <- right_join(population, migration, by = c(strata, "TrackObjects_Label_30"))

feature_migration = c("Track_Speed","Track_Speed_Y","Track_Speed_X","Track_xFMI","Track_yFMI","Track_Directionality","Track_Distance_Traveled","Track_CI")
```

```{r}
result_migration <- tibble()

seeds <- sample(1:1000, 20, replace = FALSE)

for (seed in seeds) {
  set.seed(seed)

  result_migration <- migration %>%
    filter(Metadata_matrix == "HEM") %>%
    fraction_strong(., group_by = group_by, feature_migration) %>%
    mutate(Metadata_matrix = "HEM") %>%
    mutate(feature = "migration") %>%
    mutate(seed = seed) %>%
    rbind(., result_migration)
  
  result_migration <- migration %>%
    filter(Metadata_matrix == "FN") %>%
    fraction_strong(., group_by = group_by, feature_migration) %>%
    mutate(Metadata_matrix = "FN") %>%
    mutate(feature = "migration") %>%
    mutate(seed = seed) %>%
    rbind(., result_migration)
  
    feature_var <-
    colnames(df) %>%
    str_subset("^Area|^Texture|^Track") 
    feature_var = feature_var[!feature_var == "TrackObjects_Label_30"]
    
  
  result_migration <- population_2 %>%
    filter(Metadata_matrix == "HEM") %>%
    fraction_strong(., group_by = group_by, c(feature_var,feature_migration)) %>%
    mutate(Metadata_matrix = "HEM") %>%
    mutate(feature = "-all-") %>%
    mutate(seed = seed) %>%
    rbind(., result_migration)
  
  result_migration <- population_2 %>%
    filter(Metadata_matrix == "FN") %>%
    fraction_strong(., group_by = group_by, c(feature_var,feature_migration)) %>%
    mutate(Metadata_matrix = "FN") %>%
    mutate(feature = "-all-") %>%
    mutate(seed = seed) %>%
    rbind(., result_migration)
}
```

```{r}
result_migration %>% 
  group_by(feature,Metadata_matrix) %>%
  select(-seed) %>%
  summarise_each(funs(mean)) %>%
  knitr::kable()
```
