==> 2017_07_cp_analysis/05_correlate.Rmd <==
---
title: "load_data"
author: "TB"
date: "5/15/2017"
output: html_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
library("tidyverse")
library("magrittr")
library("knitr")
library("corrplot")
library("stringr")
library("heatmaply")
```

# Input 
Define the input folders and the csv files used to load the single cell data generated by cellprofiler.
```{r}
analysis_folder <- '../../../analysis/2017_morph_gradient_3/'
csv_normalized_features <- 'normalized_features.csv'
csv_collapsed_features <- 'collapsed_features.csv'
csv_normalized_migration <- 'normalized_migration.csv'
```

# Metadata 
```{r}
strata <- c('Metadata_condition','Metadata_dose','Metadata_id','Metadata_date','Metadata_matrix')
```

# load data

## tracks
we remove an na column (feature AreaShape_EulerNumber)
```{r}
normalized <- read_csv(file.path(analysis_folder, csv_normalized_features) ) %>%
  select(-TrackObjects_Label_30, -TrackObjects_ParentObjectNumber_30) %>%
  select(-TrackObjects_TrajectoryX_30, -TrackObjects_TrajectoryY_30)
sum(complete.cases(normalized) == TRUE)
```

# normalize features 

## select Area, Texture and Track features 
```{r}
feature_var <-
  colnames(normalized) %>%
  str_subset("^Area|^Texture|^Track")
```

## collapse features to experiments
```{r}
aggregated <- cytominer::aggregate(
    population = normalized,
    variables = feature_var,
    strata = strata,
    operation = "median"
  ) %>% collect()

```

# update row names and remove vehicle (they have zero mean in all features and no correlation can be calculated)
```{r}
df <- aggregated %>% 
  filter(!Metadata_condition == "vehicle") %>%
  filter(Metadata_matrix == "HEM") %>%
  arrange(Metadata_condition) %>% 
  as.data.frame()

r_names <- df %>% 
  mutate(id = str_c(Metadata_id, Metadata_matrix, Metadata_condition, Metadata_dose, sep = ' ')) %>%
  select(id)

row.names(df) <- r_names$id

c_matrix <- df %>%
  ungroup() %>%
  select_(.dots = feature_var) %>%
  t %>%
  cor(., method = "kendall") 

corrplot(c_matrix, order = "hclust")

```
# Cluster :) 
```{r}
clusters <- df %>% 
  select_(.dots = feature_var) %>% 
  dist() %>%
  hclust()

plot(clusters)
```
# both
```{r}
heatmaply_cor(c_matrix, margins = c(100, 100, 100, 100),
          k_col = 5, k_row = 4,
          limits = c(-1,1))
```
#
